requirements.txt
python-telegram-bot[job-queue]
aiohttp
requests
python-dotenv